External Data Loads progress

+++ dataProvider/FileManager.php
@@ -11,19 +11,33 @@ if(!isset($_SESSION)) {
 	session_start();
 	session_cache_limiter('private');
 }
+$_SESSION['site']['flops'] = 0;
 class FileManager
 {
+	public $workingDir;
+	public $workingDirName;
 	public $tempDir;
 	public $fileName;
 	public $fileExtension;
+	public $fileError = '';
 	public  $src;
 
 	function __construct()
 	{
 		$this->db = new dbHelper();
-		$this->setTempDir();
-		$this->setTempDirAvailableName();
+		$this->tempDir = $_SESSION['site']['root'] . '/temp/';
+
+		//$this->setTempDir();
+		//$this->setTempDirAvailableName();
 		return;
+	}
+
+	public function cleanUp()
+	{
+
+		if(is_dir($this->workingDir)){
+			$this->deleteWorkingDir();
+		}
 	}
 
 	public function moveUploadedFileToTempDir($file)
@@ -49,33 +63,42 @@ class FileManager
 	public function extractUploadedFileToTempDir($file)
 	{
 		$this->setSrc();
-		if($this->extractFileToTemp($file['filePath']['tmp_name'])) {
+		if($this->extractFileToTempDir($file['filePath']['tmp_name'])) {
 			return true;
 		} else {
 			return false;
 		}
 	}
 
-	public function extractFileToTemp($fileSrc)
+	public function extractFileToTempDir($fileSrc)
 	{
 		$zip = new ZipArchive();
-		if ($zip->open($fileSrc) === TRUE) {
-			$zip->extractTo($this->tempDir);
-			$zip->close();
-			$this->deleteFileBySrc($this->src);
-			return true;
+		if($this->setWorkingDir()){
+			if ($zip->open($fileSrc) === TRUE) {
+				if (!($zip->extractTo($this->workingDir))) {
+			        return false;
+			    }
+				$zip->close();
+				return $this->workingDir;
+			}else{
+				return false;
+			}
 		}else{
+			$this->fileError = 'Could not create working directory';
 			return false;
 		}
+
 	}
 
-	public function extractFileToDir($fileSrc, $dir)
+	public function extractFileToDir($fileSrc, $dir, $deleteSrcFile = false)
 	{
 		$zip = new ZipArchive();
 		if ($zip->open($fileSrc) === TRUE) {
 			$zip->extractTo($dir);
 			$zip->close();
-			$this->deleteFileBySrc($fileSrc);
+			if($deleteSrcFile){
+				$this->deleteFileBySrc($fileSrc);
+			}
 			return true;
 		}else{
 			return false;
@@ -84,26 +107,31 @@ class FileManager
 
 
 
-	private function setTempDir()
+	private function setWorkingDir()
 	{
-		$tempDir = $_SESSION['site']['root'] . '/temp/';
-		if(is_dir($tempDir) || mkdir($tempDir, 0777, true)) {
-			chmod($tempDir, 0777);
-			$this->tempDir = $tempDir;
+		$workingDir = $_SESSION['site']['root'] . '/temp/'. $this->getTempDirAvailableName();
+		if(is_dir($workingDir) || mkdir($workingDir, 0777, true)) {
+			chmod($workingDir, 0777);
+			$this->workingDir = $workingDir;
 			return true;
 		} else {
 			return false;
 		}
 	}
 
-	private function setTempDirAvailableName()
+	private function getTempDirAvailableName()
 	{
 		$name = time();
 		while(file_exists($this->tempDir . $name)) {
 			$name = time();
 		}
-		$this->fileName = $name;
-		return;
+		$this->workingDirName = $name;
+		return $this->workingDirName;
+	}
+
+	public function getWorkingDirName()
+	{
+		return $this->workingDirName;
 	}
 
 	private function setFileExtensionFromFile($fileName)
@@ -118,17 +146,31 @@ class FileManager
 		return;
 	}
 
-	private function deleteTempFile()
+	private function deleteWorkingDir()
 	{
-		return unlink($this->src);
+		return $this->rmdir_recursive($this->workingDir);
 	}
 
 	public function deleteFileBySrc($src)
 	{
-		return unlink($src);
+		return $this->rmdir_recursive($src);
 	}
 
-
+	public function rmdir_recursive($dir) {
+	    $files = scandir($dir);
+	    array_shift($files);    // remove '.' from array
+	    array_shift($files);    // remove '..' from array
+	    foreach ($files as $file) {
+	        $file = $dir . '/' . $file;
+	        if (is_dir($file)) {
+	            $this->rmdir_recursive($file);
+		    continue;
+	        }
+	        unlink($file);
+	    }
+	    rmdir($dir);
+		return true;
+	}
 }
